apps: load_fw: Separate memory image store from main app
Separate memory based image store operation implementation
from main application.
